telegram_invite_bot/requirements.txt
# Основные зависимости для бота-приглашающего
pyrogram==2.0.106
python-telegram-bot==20.7
python-dotenv==1.0.0

# Асинхронные операции
asyncio-throttle==1.0.2

# Логирование и утилиты
colorlog==6.7.0

# Для работы с JSON и конфигурацией
pydantic==2.5.0

# Дополнительные утилиты
tqdm==4.66.1